+++ util/options.cc
@@ -592,8 +592,8 @@ ReadOptions::ReadOptions()
       pin_data(false),
       background_purge_on_iterator_cleanup(false),
       readahead_size(0),
-      ignore_range_deletions(false) {
-}
+      ignore_range_deletions(false),
+      max_skippable_internal_keys(0) {}
 
 ReadOptions::ReadOptions(bool cksum, bool cache)
     : verify_checksums(cksum),
@@ -608,7 +608,7 @@ ReadOptions::ReadOptions(bool cksum, bool cache)
       pin_data(false),
       background_purge_on_iterator_cleanup(false),
       readahead_size(0),
-      ignore_range_deletions(false) {
-}
+      ignore_range_deletions(false),
+      max_skippable_internal_keys(0) {}
 
 }  // namespace rocksdb